backend/requirements.txt
# Backend requirements
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic[email]==2.5.2
openai==1.3.7
python-multipart==0.0.6
requests==2.31.0
httpx==0.25.2
